requirements.txt
langchain>=0.2,<0.3
langchain-groq>=0.1,<0.2
gradio>=4.38,<5.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
